+++ crypto_kem/sntrup653/avx2/crypto_decode_653x1541.c
@@ -10,7 +10,7 @@ static inline int16 mullo(int16 x, int16 y) {
 }
 
 static inline int16 mulhi(int16 x, int16 y) {
-    return (x * (int32)y) >> 16;
+    return (int16) ((x * (int32)y) >> 16);
 }
 
 static inline __m256i add(__m256i x, __m256i y) {
@@ -79,7 +79,7 @@ void PQCLEAN_SNTRUP653_AVX2_crypto_decode_653x1541(void *v, const unsigned char 
     a0 += s[1 * i + 0]; /* -39...290 */
     a0 = mulhi(a0, 3) - mulhi(mullo(a0, -923), 71); /* -36...35 */
     a0 += (a0 >> 15) & 71; /* 0...70 */
-    a1 = (a2 << 8) + s[i] - a0;
+    a1 = (int16) ((a2 << 8) + s[i] - a0);
     a1 = mullo(a1, -22153);
 
     /* invalid inputs might need reduction mod 9402 */
@@ -100,7 +100,7 @@ void PQCLEAN_SNTRUP653_AVX2_crypto_decode_653x1541(void *v, const unsigned char 
         a0 += s[1 * i + 0]; /* -67...325 */
         a0 = mulhi(a0, 10) - mulhi(mullo(a0, -489), 134); /* -68...67 */
         a0 += (a0 >> 15) & 134; /* 0...133 */
-        a1 = (a2 << 7) + ((s[i] - a0) >> 1);
+        a1 = (int16) ((a2 << 7) + ((s[i] - a0) >> 1));
         a1 = mullo(a1, 19563);
 
         /* invalid inputs might need reduction mod 134 */
@@ -119,7 +119,7 @@ void PQCLEAN_SNTRUP653_AVX2_crypto_decode_653x1541(void *v, const unsigned char 
     a0 = mulhi(a0, 1223) - mulhi(mullo(a0, -5681), 2953); /* -1477...1782 */
     a0 += s[1 * i + 0]; /* -1477...2037 */
     a0 += (a0 >> 15) & 2953; /* 0...2952 */
-    a1 = (a2 << 8) + s[i] - a0;
+    a1 = (int16) ((a2 << 8) + s[i] - a0);
     a1 = mullo(a1, -9543);
 
     /* invalid inputs might need reduction mod 815 */
@@ -136,7 +136,7 @@ void PQCLEAN_SNTRUP653_AVX2_crypto_decode_653x1541(void *v, const unsigned char 
         a0 = mulhi(a0, 1223) - mulhi(mullo(a0, -5681), 2953); /* -1505...1514 */
         a0 += s[2 * i + 0]; /* -1505...1769 */
         a0 += (a0 >> 15) & 2953; /* 0...2952 */
-        a1 = (s[2 * i + 1] << 8) + s[2 * i] - a0;
+        a1 = (int16) ((s[2 * i + 1] << 8) + s[2 * i] - a0);
         a1 = mullo(a1, -9543);
 
         /* invalid inputs might need reduction mod 2953 */
@@ -158,7 +158,7 @@ void PQCLEAN_SNTRUP653_AVX2_crypto_decode_653x1541(void *v, const unsigned char 
         a0 = mulhi(a0, 1756) - mulhi(mullo(a0, -1206), 13910); /* -7142...7159 */
         a0 += s[2 * i + 0]; /* -7142...7414 */
         a0 += (a0 >> 15) & 13910; /* 0...13909 */
-        a1 = (a2 << 15) + (s[2 * i + 1] << 7) + ((s[2 * i] - a0) >> 1);
+        a1 = (int16) ((a2 << 15) + (s[2 * i + 1] << 7) + ((s[2 * i] - a0) >> 1));
         a1 = mullo(a1, -13437);
 
         /* invalid inputs might need reduction mod 13910 */
@@ -178,7 +178,7 @@ void PQCLEAN_SNTRUP653_AVX2_crypto_decode_653x1541(void *v, const unsigned char 
         a0 = mulhi(a0, -101) - mulhi(mullo(a0, -8891), 1887); /* -969...943 */
         a0 += s[1 * i + 0]; /* -969...1198 */
         a0 += (a0 >> 15) & 1887; /* 0...1886 */
-        a1 = (a2 << 8) + s[i] - a0;
+        a1 = (int16) ((a2 << 8) + s[i] - a0);
         a1 = mullo(a1, 5279);
 
         /* invalid inputs might need reduction mod 1887 */
@@ -234,7 +234,7 @@ void PQCLEAN_SNTRUP653_AVX2_crypto_decode_653x1541(void *v, const unsigned char 
     a0 = mulhi(a0, 2401) - mulhi(mullo(a0, -2487), 6745); /* -3497...3527 */
     a0 += s[2 * i + 0]; /* -3497...3782 */
     a0 += (a0 >> 15) & 6745; /* 0...6744 */
-    a1 = (s[2 * i + 1] << 8) + s[2 * i] - a0;
+    a1 = (int16) ((s[2 * i + 1] << 8) + s[2 * i] - a0);
     a1 = mullo(a1, -29207);
 
     /* invalid inputs might need reduction mod 7910 */
@@ -287,7 +287,7 @@ void PQCLEAN_SNTRUP653_AVX2_crypto_decode_653x1541(void *v, const unsigned char 
     a0 = mulhi(a0, 64) - mulhi(mullo(a0, -12768), 1314); /* -657...673 */
     a0 += s[1 * i + 0]; /* -657...928 */
     a0 += (a0 >> 15) & 1314; /* 0...1313 */
-    a1 = (a2 << 7) + ((s[i] - a0) >> 1);
+    a1 = (int16) ((a2 << 7) + ((s[i] - a0) >> 1));
     a1 = mullo(a1, -399);
 
     /* invalid inputs might need reduction mod 1541 */
@@ -369,7 +369,7 @@ void PQCLEAN_SNTRUP653_AVX2_crypto_decode_653x1541(void *v, const unsigned char 
 
     /* R1 ------> R0: reconstruct mod 653*[1541] */
 
-    R0[652] = 3 * R1[326] - 2310;
+    R0[652] = (int16) (3 * R1[326] - 2310);
     s -= 326;
     i = 310;
     for (;;) {
