+++ crypto_kem/sntrup761/clean/crypto_encode_761x3.c
@@ -11,9 +11,9 @@ void PQCLEAN_SNTRUP761_CLEAN_crypto_encode_761x3(unsigned char *s, const void *v
 
     for (i = 0; i < p / 4; ++i) {
         x = *f++ + 1;
-        x += (*f++ + 1) << 2;
-        x += (*f++ + 1) << 4;
-        x += (*f++ + 1) << 6;
+        x += (uint8) ((*f++ + 1) << 2);
+        x += (uint8) ((*f++ + 1) << 4);
+        x += (uint8) ((*f++ + 1) << 6);
         *s++ = x;
     }
     x = *f++ + 1;
